+++ CHANGELOG.md
@@ -170,3 +170,4 @@
 - 2023-11-13: Valid follow up, campaigns, task, find email, +5
 - 2023-11-14: Features follow up, qualify lead, sender, mod, +4
 - 2023-11-15: Init find email, LICENSE, sender, mod, +2
+- 2023-11-16: Add follow up, sender, users, deal, +3 types